많은 리소스를 동시접근하는 케이스의 성능 개선

+++ AAES/AAESResource.cs
@@ -11,8 +11,8 @@ namespace AAES
 {
     public sealed class AAESResource
     {
-        private static int LastId;
-        public int Id { get; } = Interlocked.Increment(ref LastId);
+        private static long LastId;
+        public long Id { get; } = Interlocked.Increment(ref LastId);
         public DebugInformation? DebugInfo { get; }
 
         private static readonly Task Locked = new(() => throw new InvalidOperationException());
